#315 move zocalo to external interaction folder
